Flask-Backend/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.2
cachelib==0.9.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
Flask-WTF==1.1.1
gevent==23.9.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3.post1
redis==4.6.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
zope.event==5.0
zope.interface==6.0
requests==2.31.0

